backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pandas>=2.0.0
requests>=2.31.0
twilio>=8.10.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
jinja2>=3.1.0
websockets>=12.0
python-csv>=1.0
phonenumbers>=8.13.0
